+++ packages/logger/src/index.ts
@@ -1,4 +1,4 @@
-export * from './Logger';
+export * from './logger';
 export * from './writers';
 /** @deprecated Please migrate to using the './Logger' instead. */
-export * from './LegacyLogger';
+export * from './legacyLogger';
